Make the licensing text more precise
This makes it more clear exactly what licenses
are allowed, by saying "OSI Open Source Definition
and the FSF Free Software Definition" instead of a
vague statement about OSI and FSF (which is too vague).

I included specific examples, because that should immediately
answer questions for most people (so they don't have to look up
anything). I added a note about releasing to the public domain,
which includes CC0 and software written by US government employees.
OSI won't just say CC0 is okay, in part because it's a partial
license and it's possible for a program to be released under CC0
while the releaser holds on to patents. But it *does* meet the
definition in the usual case where there are no patents being
withheld.

This text is based on:
https://www.bestpractices.dev/en/criteria/0?details=true&rationale=true#0.floss_license

However, that text also allows Debian and Fedora to decide.
Others seemed uncomfortable with that, and Eddie seemed to
want this text to look like this instead of like the best practices
badge text, so here's an attempt to address that.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>