Modify 'direct' to deal with decentralized repos
The current baseline text unintentionally presumes that a VCS must
work centrally. That is, it assumes that contributors often
share write access to the shared repo.

Many projects (like the Linux kernel) do NOT operate this way.
In more decentralized models, changes are proposed in a separate
repo. Merging those proposed changes is a separate request
to the main repo (and its primary branch), which may or may not
be granted. This still counters unintentional changes, and meets
the underlying goal, but in a different way
than the baseline text currently requires.

This attempts to modify the baseline so that it no longer

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>