Add criteria involving testing
Add criteria for automated tests to verify that a changed application
still works. Without these, it's not possible to respond in a timely
way to vulnerability reports, either in the software or its
dependencies, because there's no timely way to verify that the software
continues to work correctly.

This is intentionally mirroring OpenSSF Best Practices Badge criteria
"test" and "test_policy".

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>