Clarify OSPS-BR-01 further (#150)
* Clarify OSPS-BR-01 further

This attempts to clarify OSPS-BR-01 further.

In particular, the term "arbitrary" currently used in the criterion
doesn't really mean anything. Any sequence of bits could be
"arbitrary". It doesn't matter if the bits are arbitrary or not.

What matters is that the undesired inputs are *untrusted*.
In fact, the *rationale* uses the term "untrusted",
but the criterion and details aren't consistent with the rationale.
I think that was what was meant anyway :-).

Also: using "code execution" as a privileged resource is a
terrible example. Any CI/CD pipeline does code execution, so
expressly listing it here *confuses* things.
Let's instead list "secret exfiltration" and "final release"
as privileged operations. Those are *much* clearer as
examples of things you should be concerned about :-).

Not everyone has a separate "final release" step, and instead
consider "merge to main" the same as final release. But even in that
case, "merge to main" is a privileged operation that
is not granted to everyone, so I think it still works.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Update baseline/OSPS-BR.yaml

Co-authored-by: Evan Anderson <evan.k.anderson@gmail.com>
Signed-off-by: CRob <69357996+SecurityCRob@users.noreply.github.com>

---------

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>
Signed-off-by: CRob <69357996+SecurityCRob@users.noreply.github.com>
Co-authored-by: CRob <69357996+SecurityCRob@users.noreply.github.com>
Co-authored-by: Evan Anderson <evan.k.anderson@gmail.com>