fix: render the framework mapping references in docs (#311)
Visiting https://baseline.openssf.org/versions/devel#osps-ac-0101
shows that the content under the heading `External Framework Mappings`
is missing from the control content in the page.

This change adjusts the field naming of the baseline controls
in the data files to match the expected name of `guideline-mappings`
and adjusts the template file slightly to match the expected sci/layer2
types and fields available for templating.

Signed-off-by: Travis Truman <trumant@gmail.com>