Map Baseline controls to Software Security Code of Practice (#309)
* docs: provide more details on the frameworks mapped

This change adds descriptions for all of the frameworks and
modifies the documentation template to display the framework
details in a table.

Signed-off-by: Travis Truman <trumant@gmail.com>

* map Baseline controls to Software Security Code of Practice

This PR maps the [Software Security Code of Practice](https://www.ncsc.gov.uk/guidance/software-security-code-of-practice-assurance-principles-claims) controls to the relevant Baseline controls.

This closes https://github.com/ossf/security-baseline/issues/305
Signed-off-by: Travis Truman <trumant@gmail.com>

---------

Signed-off-by: Travis Truman <trumant@gmail.com>